+++ src/app/services/consultaCorte.ts
@@ -3,7 +3,7 @@ import { Observable, throwError } from 'rxjs';
 import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 import { map, catchError } from 'rxjs/operators';
 import { Router } from '@angular/router';
-import { cortes } from '../interfaces/cortes';
+import { cortesPlaneacion } from '../interfaces/cortes';
 
 import { AppsettingsComponent } from '../app-settings/appsettings.component'
 import { DatosTalon } from '../interfaces/datosTalon';
@@ -30,17 +30,15 @@ export class consultaCorteService {
   }
 
   getAllCortes(filtro:any){
-    return this.http.post<cortes[]>(this.appsettings.API_ENDPOINT + `planeacion/logistica/corte/getAll/`,filtro);
+    return this.http.post<cortesPlaneacion[]>(this.appsettings.API_ENDPOINT + `planeacion/logistica/corte/getAll/`,filtro);
   }
   getCortesOficinas(filtro:any){
-    console.log(this.appsettings.API_ENDPOINT + `planeacion/logistica/corte/getByIdOficina`,filtro)
-    return this.http.post<cortes[]>(this.appsettings.API_ENDPOINT + `planeacion/logistica/corte/getByIdOficina/`,filtro);
+    return this.http.post<cortesPlaneacion[]>(this.appsettings.API_ENDPOINT + `planeacion/logistica/corte/getByIdOficina/`,filtro);
   }
 
   getFindbyCorte(corte:string){
     const corteDetalle=corte.replace(/"/g, '');
-    console.log(this.appsettings.API_ENDPOINT + `planeacion/logistica/corte/getByIdCorte/`+corteDetalle);
-    return this.http.get<cortes[]>(this.appsettings.API_ENDPOINT + `planeacion/logistica/corte/getByIdCorte/`+corteDetalle);
+    return this.http.get<cortesPlaneacion[]>(this.appsettings.API_ENDPOINT + `planeacion/logistica/corte/getByIdCorte/`+corteDetalle);
   }