+++ src/app/systems/inicio/inicio-sidenav/inicio.component.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
@@ -15,6 +15,7 @@ import { MatMenuModule } from '@angular/material/menu';
 import { AuthService } from 'src/app/authentication/login/auth.service';
 import { Router } from '@angular/router';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import * as moment from 'moment';
 
 @Component({
   selector: 'app-inicio',
@@ -25,7 +26,7 @@ import { MatSnackBar } from '@angular/material/snack-bar';
     , InicioCastoresComponent, NgFor, MatMenuModule, InicioRoutingModule,]
 })
 
-export class InicioComponent {
+export class InicioComponent implements OnInit{
   @ViewChild('panel1') panel1!: MatExpansionPanel;
   @ViewChild('panel2') panel2!: MatExpansionPanel;
 
@@ -34,12 +35,26 @@ export class InicioComponent {
   @ViewChild('sidenav') sidenav!: MatSidenav;
   panelOpenState = false;
   expansionPanels: any;
-  public nombreCompletoUsuario: string;
+  public nombreCompletoUsuario!: string;
+  idOficinaActual!:string;
+  fechaActual!:string;
 
   constructor(private authService: AuthService, private router: Router, public snackBar: MatSnackBar) {
-    this.nombreCompletoUsuario = this.obtenerNombre();
   }
 
+
+  ngOnInit(): void {
+    this.nombreCompletoUsuario = this.obtenerNombre();
+      const today = new Date();
+      const fechaIniciaFormato =  moment(today).format('YYYY-MM-DD');
+      this.fechaActual = fechaIniciaFormato;
+      this.idOficinaActual = this.obtenerIdOficina() === '1100' ? '1100' : this.obtenerIdOficina();
+  }
+
+  obtenerIdOficina(): string {
+    let idoficinaJson = JSON.parse(sessionStorage.getItem('usuario')!);
+    return idoficinaJson.claveOficina;
+  }
   /**
     * openSnackBar: Funcion para ver los mensajes.
     *