+++ CHANGELOG.md
@@ -1,8 +1,22 @@
 # Planeacion Logistica
 ## Versión: 1.0.2.0
+- __Ticket/Proyecto:__ Planeacion Logistica
+- __Autor:__ Oswaldo Ramirez
+- __Fecha:__ 16/08/2023
+- __Descripción:__
+ - Se crea nuevo modulo flete optimo.
+## Versión: 1.0.1.0
 
 - __Ticket/Proyecto:__ Planeacion Logistica
-- __Author:__ Oswaldo Ramirez
+- __Autor:__ Oswaldo Ramirez
+- __Fecha:__ 16/08/2023
+- __Descripción:__
+ - Se Agregan nuevos modulos Planeacion principal y Consulta corte.
+ 
+## Versión: 1.0.0.0
+
+- __Ticket/Proyecto:__ Planeacion Logistica
+- __Autor:__ Oswaldo Ramirez
 - __Fecha:__ 22/08/2023
 - __Descripción:__
  - Se crea nuevo modulo flete optimo